refactor: Pricing and NoPricing components into Pricing folder
