+++ wealth-sync/src/app/page.tsx
@@ -2,10 +2,10 @@ import Footer from "@/components/Footer";
 import IntroductionSection from "@/components/IntroductionSection";
 import FeaturesSection from "@/components/FeaturesSection";
 import IntegrationSection from "@/components/IntegrationPlatformsSection/IntegrationSection";
-// import PricingSection from "@/components/PricingSection";
+// import PricingSection from "@/components/Pricing/PricingSection";
 import ChartPrievewSection from "@/components/ChartPrievewSection";
 import ActionSection from "@/components/ActionSection";
-import { NoPricingSection } from "@/components/NoPiricngSection/NoPricingSection";
+import { NoPricingSection } from "@/components/Pricing/NoPricingSection/NoPricingSection";
 
 export default function HomePage() {
   return (
