+++ wealth-sync/src/app/pricing/page.tsx
@@ -1,5 +1,5 @@
-import { NoPricingSection } from "@/components/NoPiricngSection/NoPricingSection";
-// import PricingSection from '@/components/PricingSection'
+// import PricingSection from '@/components/Pricing/PricingSection'
+import { NoPricingSection } from "@/components/Pricing/NoPricingSection/NoPricingSection";
 import React from "react";
 
 const PricingPage = () => {
