+++ wealth-sync/src/components/Pricing/NoPricingSection/FairPricingCard.tsx
