+++ wealth-sync/src/components/Pricing/NoPricingSection/NoPricingSection.tsx
@@ -4,7 +4,7 @@ import BuildingCommunityCard from "./BuildingCommunityCard";
 import PerfectingFeaturesCard from "./PerfectingFeaturesCard";
 import FairPricingCard from "./FairPricingCard";
 import TimeLine from "./TimeLine";
-import ShareFeedbackCard from "../ShareFeedbackCard";
+import ShareFeedbackCard from "@/components/ShareFeedbackCard";
 
 export function NoPricingSection() {
   return (
